Tests/LibWeb/Text/expected/SVG/svg-pattern-element-attributes.txt

Element found: true
id: pattern-element
x: 10
y: 20
width: 30
height: 40
